[lexical-playground] Refactor: move importers into each feature extension
Co-locate every playground node importer with the feature extension
that owns the node, and migrate the remaining React-shaped plugins to
extensions so each registered node has an extension carrying its
importer:

- Rename PollPlugin, EquationsPlugin, LayoutPlugin, MentionsPlugin,
  ExcalidrawPlugin to *Extension and turn each one into a
  defineExtension (with nodes, commands, transforms, and rule). Plugins
  whose work was only registerCommand-via-useEffect (Poll, Equations,
  Layout) are gone from Editor.tsx entirely; Mentions and Excalidraw
  keep their React UI but the node + rule live on the extension.
- Move every per-node import rule (Page-break, Tweet, YouTube, Image,
  Equation, DateTime, Layout, Mention, Excalidraw, Poll, Collapsible)
  out of the node files and into the matching feature extension's
  configExtension(DOMImportExtension, {rules: ...}). Node files no
  longer import defineImportRule.
- Simplify the rules with the new API:
  - Use regex captures (Tweet/YouTube) and selector predicates
    (DateTime's Google-Docs branch keys off a regex on data-rich-links)
    so $convertXxxElement helpers and explicit $next() escape hatches
    go away.
  - Rewrite the <details> importer on top of BlockSchema +
    ImportChildrenOpts.$onChild: BlockSchema wraps inline runs into
    paragraphs (no more hand-rolled flushPending), and $onChild siphons
    the CollapsibleTitleNode out so the rest stays in `bodyNodes` for
    the content split.
  - Inline `<details>.open` directly instead of round-tripping through
    the always-true undefined check.
- Shrink PlaygroundImportExtension to a single rule (the
  font-size/color/background-color span overlay that the legacy
  `buildHTMLConfig` carried); everything else now ships from the
  feature extensions.
- Switch the LayoutContainerNode test to use $isLayoutContainerNode +
  assert (instead of toBeInstanceOf + as cast) and depend on the new
  LayoutExtension; also point ComponentPicker, Toolbar, Editor, and
  ImageNode at the renamed paths.
