+++ packages/lexical-playground/src/nodes/ExcalidrawNode/index.tsx
@@ -17,7 +17,6 @@ import type {
 } from 'lexical';
 import type {JSX} from 'react';
 
-import {defineImportRule, sel} from '@lexical/html';
 import {DecoratorNode} from 'lexical';
 import * as React from 'react';
 
@@ -33,21 +32,6 @@ export type SerializedExcalidrawNode = Spread<
   },
   SerializedLexicalNode
 >;
-
-function $convertExcalidrawElement(el: HTMLElement): ExcalidrawNode | null {
-  const excalidrawData = el.getAttribute('data-lexical-excalidraw-json');
-  if (!excalidrawData) {
-    return null;
-  }
-  const styleAttributes = window.getComputedStyle(el);
-  const heightStr = styleAttributes.getPropertyValue('height');
-  const widthStr = styleAttributes.getPropertyValue('width');
-  const height =
-    !heightStr || heightStr === 'inherit' ? 'inherit' : parseInt(heightStr, 10);
-  const width =
-    !widthStr || widthStr === 'inherit' ? 'inherit' : parseInt(widthStr, 10);
-  return $createExcalidrawNode(excalidrawData, width, height);
-}
 
 export class ExcalidrawNode extends DecoratorNode<JSX.Element> {
   __data: string;
@@ -184,12 +168,3 @@ export function $isExcalidrawNode(
 ): node is ExcalidrawNode {
   return node instanceof ExcalidrawNode;
 }
-
-export const ExcalidrawImportRule = defineImportRule({
-  $import: (_ctx, el, $next) => {
-    const node = $convertExcalidrawElement(el);
-    return node ? [node] : $next();
-  },
-  match: sel.tag('span').attr('data-lexical-excalidraw-json', true),
-  name: '@lexical/playground/excalidraw',
-});
