+++ packages/lexical-playground/src/nodes/PlaygroundImportExtension.ts
@@ -21,22 +21,8 @@ import {
   defineExtension,
 } from 'lexical';
 
-import {CollapsibleContainerImportRule} from '../plugins/CollapsibleExtension/CollapsibleContainerNode';
-import {CollapsibleContentImportRule} from '../plugins/CollapsibleExtension/CollapsibleContentNode';
-import {CollapsibleTitleImportRule} from '../plugins/CollapsibleExtension/CollapsibleTitleNode';
 import {parseAllowedFontSize} from '../plugins/ToolbarPlugin/fontSize';
 import {parseAllowedColor} from '../ui/ColorPicker';
-import {DateTimeImportRules} from './DateTimeNode/DateTimeNode';
-import {EquationImportRules} from './EquationNode';
-import {ExcalidrawImportRule} from './ExcalidrawNode/index';
-import {ImageImportRules} from './ImageNode';
-import {LayoutContainerImportRule} from './LayoutContainerNode';
-import {LayoutItemImportRule} from './LayoutItemNode';
-import {MentionImportRule} from './MentionNode';
-import {PageBreakImportRule} from './PageBreakNode/index';
-import {PollImportRule} from './PollNode';
-import {TweetImportRule} from './TweetNode';
-import {YouTubeImportRule} from './YouTubeNode';
 
 function getPlaygroundExtraStyles(element: HTMLElement): string {
   // Parse styles from pasted input, but only if they match exactly the
@@ -110,28 +96,7 @@ const PlaygroundInlineStyleRule = defineImportRule({
  * more-specific selectors win dispatch over the generic ones (rule at
  * index 0 has the highest priority).
  */
-export const PlaygroundImportRules = [
-  // <figure type="page-break"> must precede the generic <figure> rule
-  PageBreakImportRule,
-  ...ImageImportRules,
-  ...EquationImportRules,
-  ...DateTimeImportRules,
-  ExcalidrawImportRule,
-  TweetImportRule,
-  YouTubeImportRule,
-  PollImportRule,
-  MentionImportRule,
-  LayoutContainerImportRule,
-  LayoutItemImportRule,
-  CollapsibleContainerImportRule,
-  CollapsibleContentImportRule,
-  CollapsibleTitleImportRule,
-  // The inline-style overlay matches every <span>/<b>/<i>/... and would,
-  // if it ran first, $next() into the specialized rules above and then
-  // smear playground inline styles onto their TextNode descendants.
-  // Registering it last lets those rules consume the element first.
-  PlaygroundInlineStyleRule,
-];
+export const PlaygroundImportRules = [PlaygroundInlineStyleRule];
 
 /**
  * Bundles {@link PlaygroundImportRules} into a single dependency that wires
