+++ packages/lexical-playground/src/plugins/CollapsibleExtension/CollapsibleContainerNode.ts
@@ -6,12 +6,9 @@
  *
  */
 
-import {defineImportRule, sel} from '@lexical/html';
 import {IS_CHROME, IS_FIREFOX} from '@lexical/utils';
 import {
-  $createParagraphNode,
   $getSiblingCaret,
-  $isBlockElementNode,
   $isElementNode,
   $rewindSiblingCaret,
   DOMExportOutput,
@@ -26,16 +23,6 @@ import {
   Spread,
 } from 'lexical';
 
-import {
-  $createCollapsibleContentNode,
-  $isCollapsibleContentNode,
-  CollapsibleContentNode,
-} from './CollapsibleContentNode';
-import {
-  $createCollapsibleTitleNode,
-  $isCollapsibleTitleNode,
-  CollapsibleTitleNode,
-} from './CollapsibleTitleNode';
 import {setDomHiddenUntilFound} from './CollapsibleUtils';
 
 type SerializedCollapsibleContainerNode = Spread<
@@ -44,67 +31,6 @@ type SerializedCollapsibleContainerNode = Spread<
   },
   SerializedElementNode
 >;
-
-function $assembleDetailsChildren(
-  childLexicalNodes: LexicalNode[],
-): [CollapsibleTitleNode, CollapsibleContentNode] {
-  // CollapsibleContainerNode is a shadow root that requires exactly two
-  // children: a CollapsibleTitleNode (from <summary>) followed by a
-  // CollapsibleContentNode. Arbitrary <details> markup may include loose
-  // text or block siblings; reshape the imported children into the
-  // expected structure so the editor doesn't end up with TextNodes
-  // directly under the shadow root.
-  let titleNode: CollapsibleTitleNode | null = null;
-  let contentNode: CollapsibleContentNode | null = null;
-  const bodyNodes: LexicalNode[] = [];
-  for (const child of childLexicalNodes) {
-    if (titleNode === null && $isCollapsibleTitleNode(child)) {
-      titleNode = child;
-    } else if ($isCollapsibleContentNode(child)) {
-      if (contentNode === null) {
-        // Lexical-exported markup wraps the body in a
-        // CollapsibleContentNode; reuse it instead of rebuilding.
-        contentNode = child;
-      } else {
-        // Multiple content nodes (rare): fold the extras into bodyNodes
-        // so they get appended to the canonical one below.
-        for (const grandchild of child.getChildren()) {
-          bodyNodes.push(grandchild);
-        }
-      }
-    } else {
-      bodyNodes.push(child);
-    }
-  }
-  if (titleNode === null) {
-    titleNode = $createCollapsibleTitleNode();
-  }
-  if (contentNode === null) {
-    contentNode = $createCollapsibleContentNode();
-  }
-  // CollapsibleContentNode is also a shadow root, so wrap any inline
-  // siblings in a paragraph before appending.
-  let pending: LexicalNode[] = [];
-  const flushPending = () => {
-    if (pending.length === 0) {
-      return;
-    }
-    const paragraph = $createParagraphNode();
-    paragraph.append(...pending);
-    contentNode!.append(paragraph);
-    pending = [];
-  };
-  for (const body of bodyNodes) {
-    if ($isBlockElementNode(body)) {
-      flushPending();
-      contentNode.append(body);
-    } else {
-      pending.push(body);
-    }
-  }
-  flushPending();
-  return [titleNode, contentNode];
-}
 
 export class CollapsibleContainerNode extends ElementNode {
   __open: boolean;
@@ -246,17 +172,3 @@ export function $isCollapsibleContainerNode(
 ): node is CollapsibleContainerNode {
   return node instanceof CollapsibleContainerNode;
 }
-
-export const CollapsibleContainerImportRule = defineImportRule({
-  $import: (ctx, el) => {
-    const isOpen = el.open !== undefined ? el.open : true;
-    const [titleNode, contentNode] = $assembleDetailsChildren(
-      ctx.$importChildren(el),
-    );
-    return [
-      $createCollapsibleContainerNode(isOpen).append(titleNode, contentNode),
-    ];
-  },
-  match: sel.tag('details'),
-  name: '@lexical/playground/details',
-});
