+++ packages/lexical-playground/src/plugins/ComponentPickerPlugin/index.tsx
@@ -40,12 +40,12 @@ import catTypingGif from '../../images/cat-typing.gif';
 import {EmbedConfigs} from '../AutoEmbedPlugin';
 import {INSERT_COLLAPSIBLE_COMMAND} from '../CollapsibleExtension';
 import {INSERT_DATETIME_COMMAND} from '../DateTimeExtension';
-import {InsertEquationDialog} from '../EquationsPlugin';
-import {INSERT_EXCALIDRAW_COMMAND} from '../ExcalidrawPlugin';
+import {InsertEquationDialog} from '../EquationsExtension';
+import {INSERT_EXCALIDRAW_COMMAND} from '../ExcalidrawExtension';
 import {INSERT_IMAGE_COMMAND, InsertImageDialog} from '../ImagesExtension';
-import InsertLayoutDialog from '../LayoutPlugin/InsertLayoutDialog';
+import InsertLayoutDialog from '../LayoutExtension/InsertLayoutDialog';
 import {INSERT_PAGE_BREAK} from '../PageBreakExtension';
-import {InsertPollDialog} from '../PollPlugin';
+import {InsertPollDialog} from '../PollExtension';
 import {InsertTableDialog} from '../TablePlugin';
 
 export class ComponentPickerOption extends MenuOption {
