+++ packages/lexical-playground/src/plugins/EquationsExtension/index.tsx
@@ -1,0 +1,114 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import type {JSX} from 'react';
+
+import 'katex/dist/katex.css';
+
+import {
+  CoreImportExtension,
+  defineImportRule,
+  DOMImportExtension,
+  sel,
+} from '@lexical/html';
+import {
+  $insertNodeIntoLeaf,
+  $insertNodeToNearestRoot,
+  $wrapNodeInElement,
+} from '@lexical/utils';
+import {
+  $createParagraphNode,
+  $isRootOrShadowRoot,
+  COMMAND_PRIORITY_EDITOR,
+  configExtension,
+  createCommand,
+  defineExtension,
+  LexicalCommand,
+  LexicalEditor,
+} from 'lexical';
+import {useCallback} from 'react';
+
+import {$createEquationNode, EquationNode} from '../../nodes/EquationNode';
+import KatexEquationAlterer from '../../ui/KatexEquationAlterer';
+
+type CommandPayload = {
+  equation: string;
+  inline: boolean;
+};
+
+export const INSERT_EQUATION_COMMAND: LexicalCommand<CommandPayload> =
+  createCommand('INSERT_EQUATION_COMMAND');
+
+function $convertEquationElement(el: HTMLElement) {
+  const encoded = el.getAttribute('data-lexical-equation');
+  if (!encoded) {
+    return null;
+  }
+  const equation = atob(encoded);
+  if (!equation) {
+    return null;
+  }
+  const inline = el.getAttribute('data-lexical-inline') === 'true';
+  return $createEquationNode(equation, inline);
+}
+
+const EquationImportRule = defineImportRule({
+  $import: (_ctx, el, $next) => {
+    const node = $convertEquationElement(el);
+    return node ? [node] : $next();
+  },
+  match: sel.tag('div', 'span').attr('data-lexical-equation', true),
+  name: '@lexical/playground/equation',
+});
+
+export const EquationsExtension = defineExtension({
+  dependencies: [
+    CoreImportExtension,
+    configExtension(DOMImportExtension, {rules: [EquationImportRule]}),
+  ],
+  name: '@lexical/playground/Equations',
+  nodes: [EquationNode],
+  register: editor =>
+    editor.registerCommand<CommandPayload>(
+      INSERT_EQUATION_COMMAND,
+      payload => {
+        const {equation, inline} = payload;
+        const equationNode = $createEquationNode(equation, inline);
+
+        if (inline) {
+          $insertNodeIntoLeaf(equationNode);
+          if ($isRootOrShadowRoot(equationNode.getParent())) {
+            $wrapNodeInElement(equationNode, $createParagraphNode).selectEnd();
+          }
+        } else {
+          $insertNodeToNearestRoot(equationNode);
+        }
+
+        return true;
+      },
+      COMMAND_PRIORITY_EDITOR,
+    ),
+});
+
+export function InsertEquationDialog({
+  activeEditor,
+  onClose,
+}: {
+  activeEditor: LexicalEditor;
+  onClose: () => void;
+}): JSX.Element {
+  const onEquationConfirm = useCallback(
+    (equation: string, inline: boolean) => {
+      activeEditor.dispatchCommand(INSERT_EQUATION_COMMAND, {equation, inline});
+      onClose();
+    },
+    [activeEditor, onClose],
+  );
+
+  return <KatexEquationAlterer onConfirm={onEquationConfirm} />;
+}
