+++ packages/lexical-playground/src/plugins/PageBreakExtension/index.ts
@@ -6,11 +6,18 @@
  *
  */
 
+import {
+  CoreImportExtension,
+  defineImportRule,
+  DOMImportExtension,
+  sel,
+} from '@lexical/html';
 import {$insertNodeToNearestRoot} from '@lexical/utils';
 import {
   $getSelection,
   $isRangeSelection,
   COMMAND_PRIORITY_EDITOR,
+  configExtension,
   createCommand,
   defineExtension,
   LexicalCommand,
@@ -20,7 +27,17 @@ import {$createPageBreakNode, PageBreakNode} from '../../nodes/PageBreakNode';
 
 export const INSERT_PAGE_BREAK: LexicalCommand<undefined> = createCommand();
 
+const PageBreakImportRule = defineImportRule({
+  $import: () => [$createPageBreakNode()],
+  match: sel.tag('figure').attr('type', PageBreakNode.getType()),
+  name: '@lexical/playground/page-break',
+});
+
 export const PageBreakExtension = defineExtension({
+  dependencies: [
+    CoreImportExtension,
+    configExtension(DOMImportExtension, {rules: [PageBreakImportRule]}),
+  ],
   name: '@lexical/playground/PageBreak',
   nodes: [PageBreakNode],
   register: editor =>
