+++ packages/lexical-playground/src/plugins/TwitterExtension/index.ts
@@ -6,9 +6,16 @@
  *
  */
 
+import {
+  CoreImportExtension,
+  defineImportRule,
+  DOMImportExtension,
+  sel,
+} from '@lexical/html';
 import {$insertNodeToNearestRoot} from '@lexical/utils';
 import {
   COMMAND_PRIORITY_EDITOR,
+  configExtension,
   createCommand,
   defineExtension,
   LexicalCommand,
@@ -20,7 +27,17 @@ export const INSERT_TWEET_COMMAND: LexicalCommand<string> = createCommand(
   'INSERT_TWEET_COMMAND',
 );
 
+const TweetImportRule = defineImportRule({
+  $import: ctx => [$createTweetNode(ctx.captures.id[0])],
+  match: sel.tag('div').attr('data-lexical-tweet-id', /^.+$/, {capture: 'id'}),
+  name: '@lexical/playground/tweet',
+});
+
 export const TwitterExtension = defineExtension({
+  dependencies: [
+    CoreImportExtension,
+    configExtension(DOMImportExtension, {rules: [TweetImportRule]}),
+  ],
   name: '@lexical/playground/Twitter',
   nodes: [TweetNode],
   register: editor =>
