+++ packages/lexical-playground/src/plugins/YouTubeExtension/index.ts
@@ -6,9 +6,16 @@
  *
  */
 
+import {
+  CoreImportExtension,
+  defineImportRule,
+  DOMImportExtension,
+  sel,
+} from '@lexical/html';
 import {$insertNodeToNearestRoot} from '@lexical/utils';
 import {
   COMMAND_PRIORITY_EDITOR,
+  configExtension,
   createCommand,
   defineExtension,
   LexicalCommand,
@@ -20,7 +27,19 @@ export const INSERT_YOUTUBE_COMMAND: LexicalCommand<string> = createCommand(
   'INSERT_YOUTUBE_COMMAND',
 );
 
+const YouTubeImportRule = defineImportRule({
+  $import: ctx => [$createYouTubeNode(ctx.captures.id[0])],
+  match: sel
+    .tag('iframe')
+    .attr('data-lexical-youtube', /^.+$/, {capture: 'id'}),
+  name: '@lexical/playground/youtube',
+});
+
 export const YouTubeExtension = defineExtension({
+  dependencies: [
+    CoreImportExtension,
+    configExtension(DOMImportExtension, {rules: [YouTubeImportRule]}),
+  ],
   name: '@lexical/playground/YouTube',
   nodes: [YouTubeNode],
   register: editor =>
